requirements.txt
pip install 
langchain 
langchain-openai 
youtube-transcript-api 
chromadb flask 
python-dotenv